+++ src/genie/entropy/gabac/gabac-seq-conf-set.cc
@@ -104,14 +104,14 @@ void GabacSeqConfSet::loadParameters(const core::parameter::ParameterSet &parame
         if (core::getDescriptor(desc.id).tokentype) {
             auto &descConfig = loadDescriptorDecoderCfg<entropy::paramcabac::DecoderTokenType>(parameterSet, desc.id);
             for (const auto &subdesc : getDescriptor(desc.id).subseqs) {
-                auto subseqCfg = descConfig.getSubsequenceCfg(subdesc.id.second);
+                auto subseqCfg = descConfig.getSubsequenceCfg((uint8_t)subdesc.id.second);
 
                 setConfAsGabac(subdesc.id, std::move(subseqCfg));
             }
         } else {
             auto &descConfig = loadDescriptorDecoderCfg<entropy::paramcabac::DecoderRegular>(parameterSet, desc.id);
             for (const auto &subdesc : getDescriptor(desc.id).subseqs) {
-                auto subseqCfg = descConfig.getSubsequenceCfg(subdesc.id.second);
+                auto subseqCfg = descConfig.getSubsequenceCfg((uint8_t)subdesc.id.second);
 
                 setConfAsGabac(subdesc.id, std::move(subseqCfg));
             }
